+++ CHANGELOG.md
@@ -1,3 +1,6 @@
-## 0.0.1
 
-* TODO: Describe initial release.
+## 0.1.0
+* add `maxSelectCount` property to SimpleAccordion
+## 0.0.1 - 0.0.5
+
+* simple accordion widget base uplaoded